API documentation: exported types and functions

Types

Orthography.OrthographicSystem — Type

An abstract type for orthographic systems.

Orthography.TokenCategory — Type

An abstract type for token categories.

Orthography.LexicalToken — Type

Category of alphabetic tokens.

Orthography.NumericToken — Type

Category of numeric tokens.

Orthography.PunctuationToken — Type

Category of punctuation tokens.

Functions

Public functions implemented for all subtypes of OrthographicSystem.

Orthography.codepoints — Function

Generic function to retrieve list of codepoints for an OrthographicSystem.

codepoints(ortho)

Implement codepoints function for SimpleAscii.

codepoints(ortho)

Orthography.tokentypes — Function

Generic function to retrieve a list of tokentypes for an OrthographicSystem.

Implement tokentypes function for SimpleAscii.

tokentypes(ortho)

Orthography.validchar — Function

True if ch appears in list of all valid characters for this orthography.

validchar(ortho, ch)

ch is a string possibly including more than one Julia Char but representing a single character in the orthographic system ortho.

Orthography.validstring — Function

True if all chars in s are valid.

validstring(ortho, s)

Orthography.tokenize — Function

Tokenize string s using the tokenizer of the given orthographic system.

The return value is a list of OrthographicTokens.

Tokenize citable node cn using the tokenizer of the given orthographic system.

tokenize(ortho, cn)

The return value is a list of pairings of a CitableNode and a token category. The citable node is citable at the level of the token.

Tokenize corpus c using the tokenizer of the given orthographic system.

tokenize(ortho, c)

The return value is a list of pairings of a CitableNode and a token category. The citable node is citable at the level of the token.

Working with text corpora:

Orthography.corpus_histo — Function

Create an ordered dictionary of text values for tokens in corpus c. Optionally, filter the results to include only tokens of a specified type.

corpus_histo(ortho, c)
corpus_histo(ortho, c, tokenType)

Other utilities

Orthography.nfkc — Function

Shorthand function to normalize string s to Unicode form NFKC.

nfkc(s)

Example implementation

Orthography.SimpleAscii — Type

An orthographic system for a basic alphabetic subset of the ASCII characater set.

Orthography.simpleAscii — Function

Construct a SimpleAscii with correct member values.

simpleAscii()