Exported types and functions

Types

Orthography.OrthographicSystem — Type

An abstract type for orthographic systems.

Orthography.TokenCategory — Type

An abstract type for token categories.

Orthography.LexicalToken — Type

Category of alphabetic tokens.

Orthography.NumericToken — Type

Category of alphabetic tokens.

Orthography.PunctuationToken — Type

Category of punctuation tokens.

Functions

Public functions implemented for all subtypes of OrthographicSystem.

Orthography.codepoints — Method

Generic function to retrieve codepoints member of an OrthographicSystem.

codepoints(ortho)

Orthography.tokentypes — Method

Generic function to retrieve tokentypes member of an OrthographicSystem.

Orthography.validchar — Function

True if ch appears in list of all valid characters for this orthography.

validchar(ortho, ch)

ch is a string possibly including more than one char but representing a single character in the orthographic system ortho.

Orthography.validstring — Function

True if all chars in s are valid.

validstring(ortho, s)

Orthography.tokenize — Function

Tokenize s using the tokenizer of the given orthographic system.

Orthography.nfkc — Function

Shorthand function to normalize string s to Unicode form NFKC.

nfkc(s)

Generic implementations

Missing docstring.

Missing docstring for GenericOrthography. Check Documenter's build log for details.

Orthography.SimpleAscii — Type

An orthographic system for a basic alphabetic subset of the ASCII characater set.

Orthography.simpleAscii — Function

Construct a SimpleAscii with correct member values.

simpleAscii()