Exported structs and functions

Types

Orthography.OrthographicSystem — Type

An abstract type for orthographic systems.

Orthography.TokenCategory — Type

An abstract type for token categories.

Orthography.LexicalToken — Type

Category of alphabetic tokens.

Orthography.NumericToken — Type

Category of alphabetic tokens.

Orthography.PunctuationToken — Type

Category of punctuation tokens.

Functions

Public functions implemented for all subtypes of OrthographicSystem.

Orthography.codepoints — Method

codepoints(ortho)

Generic function to retrieve codepoints member of an OrthographicSystem.

Orthography.tokentypes — Method

Generic function to retrieve tokentypes member of an OrthographicSystem.

Orthography.validchar — Function

validchar(ortho, ch)

True if ch appears in list of all valid characters for this orthography.

Orthography.validstring — Function

validstring(ortho, s)

True if all chars in s are valid.

Generic implementations

Orthography.GenericOrthography — Type

A generic model of an orthographic system implementation.

Orthography.SimpleAscii — Type

An orthographic system for a basic alphabetic subset of the ASCII characater set.