Internals

Orthography.collecttail — Function

collecttail(s, arr)
collecttail(s, arr, stripped)

Remove all trailing characters belong to list in arr, and collect them. This is useful to collect punctuation from a token, for example.

Orthography.trimtail — Function

Remove all trailing characters belong to list in arr, and and return the trimmed string.

trimtail(s, arr)

This is useful to remove punctuation from a token, for example.

Orthography.basicTypes — Function

Define an Array with all valid TokenCategorysystems in the SimpleAscii orthography.

basicTypes()

Orthography.asciiTokenizer — Function

Tokenize a string in SimpleAscii orthography.

asciiTokenizer(s)

Orthography.splitAsciiPunctuation — Function

Split off any trailing punctuation and return an Array of leading strim + trailing punctuation.

splitAsciiPunctuation(s)

Orthography.asciialphabetic — Function

Compose list of alphabetic characters in SimpleAscii.

asciialphabetic()

Orthography.asciinumeric — Function

Compose list of numeric characters in SimpleAscii.

asciinumeric()

Allow notation of positive or negative decimals.

Orthography.asciipunctuation — Function

Compose list of punctuation characters in SimpleAscii.

asciipunctuation()

Orthography.asciiCPs — Function

Define a string including all valid code points in the SimpleAscii orthography.

asciiCPs()

Orthography.asciitokenforstring — Function

Construct an OrthographicToken from a string.

asciitokenforstring(s)

s is a string representing a single token in the SimpleAscii orthography.