Internals

Orthography.collecttail — Function

collecttail(s, arr)
collecttail(s, arr, stripped)

Remove all trailing characters belong to list in arr, and collect them. This is useful to collect punctuation from a token, for example.

Orthography.trimtail — Method

Remove all trailing characters belong to list in arr, and and return the trimmed string.

trimtail(s, arr)

This is useful to remove punctuation from a token, for example.