API documentation: internals

Generic utility functions

Orthography.collecttail — Function

collecttail(s, arr)
collecttail(s, arr, stripped)

Remove all trailing characters belong to list in arr, and collect them. This is useful to collect punctuation from a token, for example.

Orthography.trimtail — Function

Remove all trailing characters belong to list in arr, and return the trimmed string.

trimtail(s, arr)

This is useful to remove punctuation from a token, for example.

Functions supporting the SimpleAscii implementation

Orthography.asciialphabetic — Function

Compose list of alphabetic characters in SimpleAscii.

asciialphabetic()

Orthography.asciinumeric — Function

Compose list of numeric characters in SimpleAscii.

asciinumeric()

Allow notation of positive or negative decimals.

Orthography.asciipunctuation — Function

Compose list of punctuation characters in SimpleAscii.

asciipunctuation()

Orthography.asciiwhitespace — Function

Compose list of recognized whitespace characters in SimpleAscii.

asciiwhitespace()

Orthography.asciiCPs — Function

Define a string including all valid code points in the SimpleAscii orthography.

asciiCPs()

Orthography.basicTypes — Function

Define an Array with all valid TokenCategorysystems in the SimpleAscii orthography.

basicTypes()

Orthography.splitAsciiPunctuation — Function

Split off any trailing punctuation and return an Array of leading strim + trailing punctuation.

splitAsciiPunctuation(s)

Orthography.asciitokenforstring — Function

Construct an OrthographicToken from a string.

asciitokenforstring(s)

s is a string representing a single token in the SimpleAscii orthography.

Orthography.isAsciiAlphabetic — Function

True if all characters in s are alphabetic.

isAsciiAlphabetic(s)

Orthography.isAsciiPunctuation — Function

True if all characters in s are punctuation.

isAsciiPunctuation(s)

Orthography.isAsciiNumeric — Function

True if all characters in s are numeric characters and at least one is a digits.

isAsciiNumeric(s)

Obviously not a real-world definition of syntax for a numeric token.